PR15315: Filter inode-uprobe processes in the handler
We already had process tracking, mainly for setting semaphores, but this
gives the list needed to check if an incoming uprobe hit is for a
process we care about.  So now stapiu_probe_prehandler will consult this
list before proceeding with the rest of the probe handler.

This also updates locking so that readers of this process list have a
more fine-grained lock, rather than the global stapiu_process_lock.
Code that needs to reserve/free a uprobe slot will both the global lock
and the specific target lock.