PR14665: improve stapdyn's signal handling
This changes stapdyn's signal handling to be a bit more like staprun's.
The signals considered now are SIGHUP, SIGINT, SIGTERM, and SIGQUIT,
with the latter being a more immediate exit.  The others will wait
around for any created process to finish, but this is debatable.

The heuristic of counting signals is gone, as it just didn't work very
well.  Instead this just tracks a mask of signals we've seen.  We also
don't try to kill the target process anymore, since staprun doesn't do
that either.  If a user hits CTRL-C on the command line, that will go
to the whole process group which includes the target.