PR15044: reversed patches in functions too popular in probe aliases
The following functions had their cutesy errors restored:
    - tapset/linux/ip.stp: format_ipaddr()
    - tapset/errno.stp: returnstr()
    - tapset/uconversions.stp: user_string()

These functions are used by many probe aliases, especially in the
*syscall* tapsets and for format_ipaddr(), in the network-related
tapsets, such as rpc.stp, tcp.stp, and netfilter.stp. Thus, it is not
ideal to throw an error in case of failure. See the discussion in the
associated bug report for more information.