Fix uconversion.stp function summaries to be a single line
The tool that extracts the information from the comments for the
tapset documentation needs the summary to be on a single line.  If the
summary continues to another line, the tool thinks the second line is
part of the description and will fail generating output because there
are multiple description for the same function.