PR15278: Make use of UPROBE_HANDLER_REMOVE
When a uprobe handler is called in a process we didn't want (thanks to
shared inodes), we already skip it if it's not in our process list.  Now
we also return UPROBE_HANDLER_REMOVE to let uprobes know the probe can
be removed from this particular process, so only participating processes
will incur breakpoint overhead.  We do have to be careful that task
finder's process coverage is complete though, so we don't remove probes
from a process that just wasn't tracked yet.

* runtime/linux/task_finder2.c (stap_task_finder_complete): New test to
  see if task_finder has hooked all processes yet; set after post_init.
* runtime/linux/uprobes-inode.c (stapiu_probe_prehandler): In processes
  which we're about to skip, use UPROBE_HANDLER_REMOVE if task finder
  says it has complete coverage.