fix clamp_t arguments in user_string_n_quoted()
The 'len' parameter passed to _stp_text_str() was incorrectly set to
1 higher than need be. The function _stp_text_str() does not count the
trailing NULL. In contrast, the function _stp_strncpy_from_user() does,
which is also used in this tapset, in which case 'len' should indeed be
incremented (see e.g. user_string_n()). This may have been the source
of the confusion.