(nd_)syscall.exp: adapt tests and fix umask test failure
The recent changes in the *syscall tapsets were causing failures in a
few of the C files tested as part of nd_syscall.exp and syscall.exp.
These files had their expected output changed to match what is now
output.

Also, umask.c had been failing from before only because a '0' is
printed, not '00' as expected.