uprobes-inode.c: add UPROBES_HITCOUNT
Added a new directive 'UPROBES_HITCOUNT' which causes stapiu to keep
track of the number of times stapiu_probe_prehandler() and
stapiu_probe_handler() are called. Upon exiting, stapiu_exit() prints
these numbers.

These are important statistics which allow us to verify that uprobe
filtering is working as it should.