Updated testsuite/systemtap.examples/process/mutex_contention.stp example.
Stats that were supposed to be collected per FUTEX_WAIT call site were
being collected per mutex init stack. So, in the "aggregate stats" section
of the report, you'd see a bunch of stacks with pthread_mutex_init at the
top, which was very confusing because a futex isn't contended when it's
created.