fix umodname()-related issues in testsuite
In vma_vdso.stp, calling umodname() alone was getting elided because of
the /* pure */ marker present. The test was modified so that the value
of umodname() is used (even though the if branch will never be taken).

In usymbols.exp and pf3.stp, wrapped umodname() and modname() calls
between try/catch blocks.

Reported-by: Mark Wielaard <mjw@redhat.com>