PR14927: parse SYSTEMTAP_COLOR on-the-fly
Colors are no longer stored in memory. Instead, upon each request to
colorize(), the SYSTEMTAP_COLOR env var is parsed and queried for the
value of the key of interest.

This allows colorize() and parse_stap_color() to be more self-contained
and easier to eventually port to other parts of SystemTap, such as
staprun/io.