SDT: Use %foo for section types rather than "foo"
GAS can take section types as "foo", @foo, or %foo.  We had the quoted
form, but this is not documented, and llvm-mc doesn't support it.  Using
@foo starts a comment on ARM, so we're left with %foo.

For note, we have to escape it to %%note for the asm()'s operand
expansion, but leave it as %note for pure assembly.  For progbits, there
are no operands, so it's %progbits either way.

Thanks to Martin C. Martin for first reporting this to LLVM[1], and then
returning to systemtap with a suggested compatibility fix.
 [1] http://llvm.org/bugs/show_bug.cgi?id=17270