testsuite: split std_asm out of std_misc.exp
The asm part of sdt_misc.exp was not really dependent on any other part
of that testcase, so we can let it stand alone for direct testing.  Its
reported name in test results is left as "sdt_misc", so it can still be
easily compared to previous test runs.