stapdyn: warn on !isInstrumentable functions
The most common reason I've found that Dyninst won't allow us to
instrument a function is an indirect jump.  This prevents them from
creating a control-flow graph, so they conservatively refuse.

Explicit indirect jumps are rare, like in the internals of longjmp, but
they may also occur for switches that are implemented with a jump table,
or tail calls to a vtable function, for instance.