PR16165: extend print_format::create
In an effort to restrict token creation, we change the
print_format::create() function to accept a string to denote the type of
print statement we want, rather than always deriving it from the token's
content. This allows us to skip the creation of tokens in a few places
that synthesize print statements.

- staptree.h: update print_format::create() declaration to accept new
  parameter and add new member print_format_type
- staptree.cxx: update print_format::create() and use print_format_type
  in print_format::print()
- elaborate.cxx: don't create token, use new parameter instead
- tapset-mark.cxx: ditto
- tapset-utrace.cxx: ditto
- tapsets.cxx: ditto