testsuite: Support prelink even on NFS paths
We had an issue that prelink would fail trying to restore selinux
context if the file is on NFS, e.g. for someone working in NFS $HOME.
It turns out prelink see nfs_t on the source, but isn't allowed to set
nfs_t on the destination (even though it will already be nfs_t).

Now we have a [prelink] proc for test to run prelink though a mktemp
file.  The source is copied to tmp, prelinked, and copied back.