Set git-describe --abbrev=12 for consistency and future-proofing
Git's default abbrev is 7, with smarts to disambiguate the SHA1 for that
given moment.  Torvalds has recommended core.abbrev = 12 for kernel
developers to help avoid future as-yet-unknown collisions.

It becomes an issue to our scripts if this setting is not deterministic.
For instance, "make && sudo make install" will run git_version.sh with
$USER's git config, then root's config, but we don't want git_version.h
to be regenerated just for that difference.

Now our scripts use an explicit git-describe --abbrev=12 to be safe.